restructure file locations
