--- frontend/src/App.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-
-class App extends Component {
-  render() {
-    return (
-        <div className="App">
-            <button type="button" className="btn btn-light">Hello World</button>
-        </div>
-    );
-  }
-}
-
-export default App;
