+++ frontend/src/components/App/__tests__/App.test.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from '..';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
