+++ frontend/src/components/App/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const App = () => (
+    <div className="App">
+        <button type="button" className="btn btn-light">Hello World</button>
+    </div>
+)
+
+export default App;
