+++ frontend/src/services/README.md
@@ -1,0 +1,1 @@
+stuff, see. https://github.com/jeffersonRibeiro/react-shopping-cart/tree/master/src/services
