+++ frontend/src/static/README.md
@@ -1,0 +1,1 @@
+Image files etc.
